# HTNest.Data/Model/Category/UpdateCategoryModel.cs
using System;
using System.Collections.Generic;
using System.ComponentModel.DataAnnotations;
using System.Linq;
using System.Text;
using System.Threading.Tasks;

namespace HTNest.Data.Model.Category
{
    public class UpdateCategoryModel
    {
        [Required]
        [MinLength(3, ErrorMessage = "CategoryName least 3 character")]
        public string? categoryName { get; set; }

    }
}