# HTNest/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "NZ Walks API",
    "version": "V1"
  },
  "paths": {
    "/api/Authen/Login": {
      "post": {
        "tags": [
          "Authen"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        }
      }
    },
    "/api/Authen/Register": {
      "post": {
        "tags": [
          "Authen"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/RegisterLoginModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
    "/api/Cart/Add-cart": {
      "post": {
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "productId",
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Cart/{id}": {
      "put": {
        "tags": [
          "Cart"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": [
          "Cart"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
    "/api/Category": {
      "get": {
        "tags": [
          "Category"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "categoryName",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Category"
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Category"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateCategoryModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
    "/api/Category/{id}": {
      "get": {
        "tags": [
          "Category"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Category"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateCategoryModel"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Category"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int32"
          }
        ],
        "responses": {
          "200": {
            "description": "Success",
            "schema": {
              "$ref": "#/definitions/Category"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "Category": {
      "type": "object",
      "properties": {
        "categoryID": {
          "format": "int32",
          "type": "integer"
        },
        "categoryName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "product": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "CreateCategoryModel": {
      "required": [
        "categoryName"
      ],
      "type": "object",
      "properties": {
        "categoryName": {
          "minLength": 3,
          "type": "string"
        }
      }
    },
...
    "Feedback": {
      "type": "object",
      "properties": {
        "feedbackId": {
          "format": "int32",
          "type": "integer"
        },
        "rating": {
          "format": "int32",
          "type": "integer"
        },
        "userName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "comment": {
          "type": "string"
        },
        "feedbackTime": {
          "format": "date-time",
          "type": "string"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      }
    },
    "LoginModel": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Order": {
      "type": "object",
      "properties": {
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "userName": {
          "type": "string"
        },
        "orderDate": {
          "format": "date-time",
          "type": "string"
        },
        "user": {
          "$ref": "#/definitions/User"
        },
        "orderDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetail"
          }
        }
      }
    },
    "OrderDetail": {
      "type": "object",
      "properties": {
        "orderDetailId": {
          "format": "int32",
          "type": "integer"
        },
        "quantity": {
          "format": "int32",
          "type": "integer"
        },
        "totalPrice": {
          "format": "double",
          "type": "number"
        },
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "orderId": {
          "format": "int32",
          "type": "integer"
        },
        "order": {
          "$ref": "#/definitions/Order"
        },
        "product": {
          "$ref": "#/definitions/Product"
        }
      }
    },
    "Product": {
      "type": "object",
      "properties": {
        "productId": {
          "format": "int32",
          "type": "integer"
        },
        "productCode": {
          "type": "string"
        },
        "image": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "stockQuantity": {
          "format": "int32",
          "type": "integer"
        },
        "categoryId": {
          "format": "int32",
          "type": "integer"
        },
        "styleId": {
          "format": "int32",
          "type": "integer"
        },
        "status": {
          "type": "string"
        },
        "price": {
          "format": "double",
          "type": "number"
        },
        "ingredient": {
          "type": "string"
        },
        "discount": {
          "format": "double",
          "type": "number"
        },
        "warning": {
          "type": "string"
        },
        "createTime": {
          "format": "date-time",
          "type": "string"
        },
        "volume": {
          "format": "double",
          "type": "number"
        },
        "feedbacks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Feedback"
          }
        },
        "category": {
          "$ref": "#/definitions/Category"
        },
        "style": {
          "$ref": "#/definitions/Style"
        },
        "orderDetail": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/OrderDetail"
          }
        }
      }
    },
    "RegisterLoginModel": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "dob": {
          "format": "date-time",
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "phoneNumber": {
          "type": "string"
        },
        "createdDate": {
          "format": "date-time",
          "type": "string"
        }
      }
    },
    "Role": {
      "type": "object",
      "properties": {
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "roleName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "user": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/User"
          }
        }
      }
    },
    "Style": {
      "type": "object",
      "properties": {
        "styleId": {
          "format": "int32",
          "type": "integer"
        },
        "styleName": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "products": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Product"
          }
        }
      }
    },
    "UpdateCategoryModel": {
      "required": [
        "categoryName"
      ],
      "type": "object",
      "properties": {
        "categoryName": {
          "minLength": 3,
          "type": "string"
        }
      }
    },
...
    "User": {
      "type": "object",
      "properties": {
        "userName": {
          "type": "string"
        },
        "fullName": {
          "type": "string"
        },
        "dob": {
          "format": "date-time",
          "type": "string"
        },
        "password": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "roleId": {
          "format": "int32",
          "type": "integer"
        },
        "phoneNumber": {
          "type": "string"
        },
        "status": {
          "type": "string"
        },
        "createDate": {
          "format": "date-time",
          "type": "string"
        },
        "role": {
          "$ref": "#/definitions/Role"
        },
        "feedbacks": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Feedback"
          }
        },
        "order": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Order"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}